% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criarBase.R
\name{criabase}
\alias{criabase}
\title{CRIAÇÃO DO DATAFRAME DOS DADOS PARA ANÁLISE}
\usage{
criabase(
  nome_dados,
  extensao,
  planilha = NULL,
  separador = NULL,
  exclusivo = NULL
)
}
\arguments{
\item{extensao}{tipo do arquivo.}

\item{separador}{separador de células de tabela.}
}
\description{
CRIAÇÃO DO DATAFRAME DOS DADOS PARA ANÁLISE
}
