% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excecao.R
\name{excecaoA}
\alias{excecaoA}
\title{Controle de exceção: Caracteres e Números}
\usage{
excecaoA(funcao, objeto, padrao, min, max, atual)
}
\arguments{
\item{objeto}{Dados da variável a ser verificada.}

\item{padrao}{Padrão caso a variável entre na exceção.}

\item{min}{Valor mínimo para variável entrar na exceção.}

\item{max}{Valor máximo para variável entrar na exceção.}

\item{atual}{Valor a ser mantido se não entrar na exceção.}
}
\description{
Controle de exceção: Caracteres e Números
}
